+++ NEWS.md
@@ -5,6 +5,7 @@
 * Remove explicit R6 dependency as imported via `gargle`
 * Enable some customization of `googleSignInUI`'s logout button via `logout_name` and `logout_class` parameters
 * Expose information whether user is signed in via `googleSignIn` module
+* Auto-auth by specifying an email address in `gar_auto_auth()`
 
 # googleAuthR v0.8.1